backend-python/requirements.txt
alembic
amqp
annotated-doc
annotated-types
anyio
bcrypt
bidict
billiard
celery
certifi
cffi
click
click-didyoumean
click-plugins
click-repl
cryptography
dnspython
dotenv
ecdsa
email-validator
fastapi
geographiclib
geopy
greenlet
h11
hiredis
httpcore
httptools
httpx
idna
kombu
Mako
MarkupSafe
packaging
passlib
pillow
prompt_toolkit
psycopg2-binary
pyasn1
pycparser
pydantic
pydantic-settings
pydantic_core
python-dateutil
python-dotenv
python-engineio
python-jose
python-multipart
python-socketio
PyYAML
qrcode
redis
rsa
simple-websocket
six
SQLAlchemy
starlette
typing-inspection
typing_extensions
tzdata
tzlocal
uvicorn
uvloop
vine
watchfiles
wcwidth
websockets
wsproto
slowapi
structlog
python-json-logger
